requirements.txt
pandas==2.1.4
openpyxl==3.1.2
google-generativeai==0.8.3
jinja2==3.1.2
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
aiofiles==23.2.0
python-dotenv==1.0.0
loguru==0.7.2
lxml==4.9.3